skin-color framework "completed"
after moving to GitHub and VS code, I figured out a system to create all of the buttons for a particular frame; in this case, skcolor_page makes ten buttons, and knows exactly where to place them based on idx. I have a feeling because I understand how to display images and how to create x number of buttons, this system will be used from now on. Also, when the user picks an option, the methods are in Cosmetic.py